# docs/swagger.yaml
basePath: /
definitions:
  handler.ChannelWithPodchannels:
    properties:
      channel:
        $ref: '#/definitions/queries.Channel'
      podchannels:
        items:
          $ref: '#/definitions/queries.Podchannel'
        type: array
    type: object
  queries.Channel:
    properties:
      created_at:
        type: string
      id:
        type: integer
      name:
        type: string
      updated_at:
        type: string
    type: object
  queries.Message:
    properties:
      author_id:
        type: string
      content:
        type: string
      created_at:
        type: string
      id:
        type: integer
      podchannel_id:
        type: integer
      updated_at:
        type: string
    type: object
  queries.Podchannel:
    properties:
      channel_id:
        type: integer
      created_at:
        type: string
      id:
        type: integer
      name:
        type: string
      types:
        type: string
      updated_at:
        type: string
    type: object
info:
  contact: {}
  description: websocket chats
  title: LilDiscord Api
  version: "1.0"
paths:
  /channel:
    get:
      consumes:
      - application/json
      description: Get one channel
      operationId: get-channel
      parameters:
      - description: ID of the channel
        in: query
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.ChannelWithPodchannels'
      summary: Get one channel
      tags:
      - Channel
  /channel/create:
    post:
      consumes:
      - application/json
      description: Create Channel
      operationId: create-channel
      parameters:
      - description: Name of Channel
        in: query
        name: name
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/queries.Channel'
      summary: Create Channel
      tags:
      - Channel
  /channels:
    get:
      consumes:
      - application/json
      description: Get channels
      operationId: get-channels
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/queries.Channel'
            type: array
      summary: Get channels
      tags:
      - Channel
  /podchannel/create:
    post:
      consumes:
      - application/json
      description: Create one podchannel
      operationId: create-podchannel
      parameters:
      - description: Name of the podchannel
        in: query
        name: name
        required: true
        type: string
      - description: type of the podchannel
        in: query
        name: types
        required: true
        type: string
      - description: channel of the podchannel
        in: query
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.ChannelWithPodchannels'
      summary: Create one podchannel
      tags:
      - PodChannel
  /podchannel/message:
    get:
      consumes:
      - application/json
      description: mess podchannel
      operationId: get-podchannel-message
      parameters:
      - description: podchannel id
        in: query
        name: podchannel_id
        required: true
        type: integer
      - description: limit
        in: query
        name: limit
        required: true
        type: integer
      - description: page
        in: query
        name: page
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/queries.Message'
            type: array
      summary: Get Messages PodChannel
      tags:
      - PodChannel
  /podchannels:
    get:
      consumes:
      - application/json
      description: Get podchannels
      operationId: get-podchannels
      parameters:
      - description: ID of the podchannel
        in: query
        name: channelId
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/queries.Podchannel'
            type: array
      summary: Get podchannels
      tags:
      - PodChannel
swagger: "2.0"